[PERF](server): AccountImage 제거
AccountImage를 제거하고 Account 필드에 ImageFile 추가
