+++ molly-spring/src/main/java/kr/co/kumoh/illdang100/mollyspring/dto/account/AccountRespDto.java
@@ -1,5 +1,6 @@
 package kr.co.kumoh.illdang100.mollyspring.dto.account;
 
+import kr.co.kumoh.illdang100.mollyspring.domain.account.Account;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -15,5 +16,16 @@ public class AccountRespDto {
         private String nickname;
         private String provider;
         private String email;
+
+        public AccountProfileResponse(Account account) {
+            this.nickname = account.getNickname();
+            this.provider = getProviderFromUsername(account.getUsername());
+            this.email = account.getEmail();
+        }
+
+        private String getProviderFromUsername(String username) {
+            int idx = username.indexOf('_');
+            return username.substring(0, idx);
+        }
     }
 }
